Added French translation file
Only "All" and the article "meta information" are translated to French (written by, posted in, etc.). The rest are mostly technical terms and labels from the backend which either cannot be properly translated or would only cause irritation in any communication about settings because labels are not consistent anymore and thus cannot be found when looking for a button with a particular English label for instance.